fixed pr
