+++ Simulator/Generator/LongLatCoordinate.py
@@ -1,0 +1,8 @@
+class LongLatCoordinate:
+    """
+    Wrapper for long-lat coordinates
+    """
+    def __init__(self, long: float, lat: float):
+        self.long = long
+        self.lat = lat
+
