+++ Simulator/Generator/__init__.py
@@ -1,1 +1,2 @@
 from .Generator import Generator
+from .LongLatCoordinate import LongLatCoordinate
